fitted 13

+++ extension/Dynamic Filaments/+DF/plotTrackSlide.m
@@ -47,7 +47,7 @@ if ~all(isnan(track.itrace(frame,:)))
     iframe = frame - 4;
     tipx = - track.Data(2:end-10,2);
     itrace = track.itrace(frame,:);
-    x = (((0:length(itrace)-1)-40)*157/4) + tipx(1);
+    x = double((((0:length(itrace)-1)-40)*157/4) + tipx(1));
     plot(x,itrace);
 %     if isnan(track.Data(frame,2))
 %         if frame+1 > size(track.Data,1)
